requirements.txt
foxglove-sdk==0.14.3
lz4==4.4.4
mcap==1.3.0
PyGObject==3.54.5
zstandard==0.25.0
